extract dataclasses for lulucf2018
